requirements.txt
streamlit>=1.33.0
opencv-python>=4.8.0.76
Pillow>=10.0.0
numpy>=1.24.0
